==> 1. clean-data.ipynb <==
# Part 1. clean data
here we will clean the text data and encode the audio. and save both in disk

import numpy as np
import pandas as pd
import datasets

train_df=pd.read_csv('csv/train_data.csv')
train_df.drop_duplicates(inplace=True)
print(len(train_df))
#dataframe should have two columns, text and audio path

test_df=pd.read_csv('csv/test_data.csv')
test_df.drop_duplicates(inplace=True)
print(len(test_df))
test_df.head()

import warnings
warnings.filterwarnings('ignore')

test_data=datasets.Dataset.from_pandas(test_df)
train_data=datasets.Dataset.from_pandas(train_df)

urdata = datasets.DatasetDict({
    'train': train_data,
    'test':test_data})
urdata

import re
import unicodedata
from urduhack.normalization import *
from urduhack.preprocessing import *
chars_to_ignore_regex = """[\!\؛\،\٫\؟\۔\٪\"\'\:\-\‘\’\`]"""
def clean_data(batch):
    text=remove_punctuation(batch["text"])
    text=normalize_whitespace(text)
    text=remove_accents(text)
    text=normalize_characters(text)
    text=normalize_combine_characters(text)
    text=remove_diacritics(text)
    text=normalize(text)
#     text = re.sub('\d+', '',text)
    text = re.sub('[a-zA-Z]+', '',text)
    text = re.sub('\u200c', '',text)
    text = re.sub('\u200f', '',text)
    text = re.sub('\ufeff', '',text)
    text = unicodedata.normalize("NFKC", text)
    text = re.sub(chars_to_ignore_regex, "", text)
    #unify some character manually
    text = re.sub('\u0623','\u0627',text)#أ
    text = re.sub('\u06C3', '\u062A',text)#ه
    text = re.sub('\u0647', '\u06C1',text)#ۃ
    text = re.sub('\u06C2', '\u0621',text)#ۂ
    text = re.sub('\u0624', '\u0648',text)#ؤ
    text = re.sub('\u06D3', '\u06D2',text)#ۓ
    text = re.sub('\u0649', '\u06cc',text)#ى
    text = re.sub('\u0626', '\u06cc',text)#ئ
    text = re.sub('\u064a', '\u06cc',text)#ي
    return {"text":text }


urdata=urdata.map(clean_data)

#saving the cleaned text, will later used in language model
text=[i['text'] for i in urdata['train']]
pd.DataFrame(text,columns=['text']).to_csv('text4lm.csv',index=False)
#will later use this test data text to evaluate our model 
text=[i['text'] for i in urdata['test']]
pd.DataFrame(text,columns=['text']).to_csv('tex-test.csv',index=False)
#both these files will be used in part 3

len(text)

def extract_all_chars(batch):
  all_text = " ".join(batch["text"])
  vocab = list(set(all_text))
  return {"vocab": [vocab], "all_text": [all_text]}

vocabs = urdata.map(extract_all_chars, batched=True, batch_size=-1, keep_in_memory=True, remove_columns=urdata.column_names["train"])

vocab_list = list(set(vocabs["train"]["vocab"][0]) | set(vocabs["test"]["vocab"][0]))
vocab_dict = {v: k for k, v in enumerate(vocab_list)}
print(len(vocab_dict))

vocab_dict["|"] = vocab_dict[" "]
del vocab_dict[" "]
vocab_dict

vocab_dict["[UNK]"] = len(vocab_dict)
vocab_dict["[PAD]"] = len(vocab_dict)
print(len(vocab_dict))

import json
with open('vocab.json', 'w') as vocab_file:
    json.dump(vocab_dict, vocab_file)

from transformers import Wav2Vec2CTCTokenizer
tokenizer = Wav2Vec2CTCTokenizer("./cache/saved/vocab.json", unk_token="[UNK]", pad_token="[PAD]", word_delimiter_token="|")

from transformers import Wav2Vec2FeatureExtractor

feature_extractor = Wav2Vec2FeatureExtractor(feature_size=1, sampling_rate=16000, padding_value=0.0, do_normalize=True, return_attention_mask=True)

from transformers import Wav2Vec2Processor

processor = Wav2Vec2Processor(feature_extractor=feature_extractor, tokenizer=tokenizer)

urdata

urdata['train'][0]

urdata = urdata.cast_column("path", datasets.Audio(sampling_rate=16_000))
urdata

import IPython.display as ipd
import numpy as np
import random

rand_int = random.randint(0, len(urdata['train'])-1)

print(urdata['train'][rand_int]["text"])
ipd.Audio(data=urdata['train'][rand_int]["path"]["array"], autoplay=True, rate=16000)

urdata.column_names

def prepare_dataset(batch):
    audio = batch["path"]
    # batched output is "un-batched"
    batch["input_values"] = processor(audio["array"], sampling_rate=audio["sampling_rate"]).input_values[0]
    batch["input_length"] = len(batch["input_values"])
    
    with processor.as_target_processor():
        batch["labels"] = processor(batch["text"]).input_ids
    return batch

urdudata = urdata.map(prepare_dataset, remove_columns=urdata.column_names["train"],num_proc=8)

urdudata.save_to_disk("cache/saved")
