==> 3. language model.ipynb <==
# Part 3. language model
Here will create Language model   
https://huggingface.co/blog/wav2vec2-with-ngram

from huggingface_hub import notebook_login
notebook_login()

!wget -O - https://kheafield.com/code/kenlm.tar.gz | tar xz

!mkdir kenlm/build && cd kenlm/build && cmake .. && make -j2

!ls

import pandas as pd
df=pd.read_csv('text4lm.csv')
df.dropna(inplace=True)
df.head()


with open("text.txt", "w") as file:
  file.write(" ".join(df.text.values))

!kenlm/build/bin/lmplz -o 5 <"text.txt" > "5gram.arpa"

!head -20 5gram.arpa

with open("5gram.arpa", "r") as read_file, open("5gram_correct.arpa", "w") as write_file:
  has_added_eos = False
  for line in read_file:
    if not has_added_eos and "ngram 1=" in line:
      count=line.strip().split("=")[-1]
      write_file.write(line.replace(f"{count}", f"{int(count)+1}"))
    elif not has_added_eos and "<s>" in line:
      write_file.write(line)
      write_file.write(line.replace("<s>", "</s>"))
      has_added_eos = True
    else:
      write_file.write(line)

from transformers import AutoProcessor

processor = AutoProcessor.from_pretrained('urdumodel',local_files_only=True)

vocab_dict = processor.tokenizer.get_vocab()
sorted_vocab_dict = {k.lower(): v for k, v in sorted(vocab_dict.items(),
                                                     key=lambda item: item[1])}
print(len(sorted_vocab_dict))
sorted_vocab_dict

#pip install https://github.com/kpu/kenlm/archive/master.zip pyctcdecode
from pyctcdecode import build_ctcdecoder

decoder = build_ctcdecoder(
    labels=list(sorted_vocab_dict.keys()),
    kenlm_model_path="5gram_correct.arpa",
)

from transformers import Wav2Vec2ProcessorWithLM

processor_with_lm = Wav2Vec2ProcessorWithLM(
    feature_extractor=processor.feature_extractor,
    tokenizer=processor.tokenizer,
    decoder=decoder
)

download the model locally 
https://discuss.huggingface.co/t/download-models-for-local-loading/1963

!ls 

from huggingface_hub import Repository
repo = Repository(local_dir="urdumodel")


processor_with_lm.save_pretrained("urdumodel")

!kenlm/build/bin/build_binary urdumodel/language_model/5gram_correct.arpa urdumodel/language_model/5gram.bin

!rm urdumodel/language_model/5gram_correct.arpa && tree -h urdumodel/

repo.push_to_hub('URDU-ASR-LM')
